requirements.txt
# Core Dependencies
requests>=2.28.0
python-dotenv>=1.0.0
rich>=13.3.0
pygame>=2.5.2

# Core LLM and AI APIs
openai>=1.30.1
cohere>=4.38
groq>=0.4.1
google-generativeai
edge-tts>=6.1.10

# GUI and Web
eel
pywhatkit>=5.4
keyboard>=0.13.5
AppOpener>=1.6
bs4
Pillow>=10.3.0

# Web Automation / Search
googlesearch-python>=1.2.3

# System & Audio Processing
pyautogui>=0.9.54
pydub>=0.25.1
pyperclip>=1.8.2

# Miscellaneous Enhancements
pyfiglet>=0.8.post1
numpy

# Optional Dependencies
# Uncomment to enable optional features
# pdfminer.six>=20221105   # PDF reading
# PyPDF2>=3.0.1            # PDF writing
# pytesseract>=0.3.10      # OCR for images

# Development Tools
black>=23.7.0
isort>=5.12.0
virtualenv>=20.24.0
gitpython>=3.1.31

# Linting / Type Checking / Tests / Docs
flake8>=6.0.0
mypy>=1.5.0
pytest>=7.2.2
pytest-cov>=4.0.0
sphinx>=7.2.5
sphinx-rtd-theme>=1.2.0

# Security
bandit>=1.7.4
